post-merge conflict resolution

+++ src/SalesManager.java
@@ -14,9 +14,9 @@ public class SalesManager {
         }
         return max;
     }
-    public int min() {
-        int min = sales[0];
-        for (int sale : sales) {
+    public long min() {
+        long min = sales[0];
+        for (long sale : sales) {
             if (sale < min) {
                 min = sale;
             }
@@ -26,7 +26,7 @@ public class SalesManager {
 
     public int truncatedMean() {
         int trMean = 0;
-        for (int sale : sales) {
+        for (long sale : sales) {
             if (sale == this.max() || sale == this.min()) {
                 continue;
             }
